requirements.txt
annotated-types
anyio
blinker
brotlicffi
certifi
cffi
charset-normalizer
click
distro
dotenv
Flask
h11
httpcore
httpx
idna
itsdangerous
Jinja2
jiter
MarkupSafe
openai
pycparser
pydantic
pydantic-core
PySocks
requests
sniffio
tqdm
typing-inspection
typing_extensions
urllib3
websockets
Werkzeug
httpx[http2]
langchain>=0.2,<0.3
langchain-openai>=0.1,<0.2
langchain-community>=0.2,<0.3
faiss-cpu>=1.8
pymilvus
mem0ai
asyncio
pillow
apscheduler